MAINT: add hyperu from xsf and remove cython
